+++ gems/aws-sdk-s3/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Expose presigned request status to the request handler stack #2513
+
 1.94.0 (2021-04-27)
 ------------------